+++ vss_cli/config.py
@@ -966,41 +966,56 @@ class Configuration(VssManager):
         )
         return rv
 
-    def get_vm_by_id_or_name(self, vm_id: str, silent=False) -> Optional[List]:
-        """Get virtual machine by identifier or name."""
+    def get_vm_by_id_or_name(
+        self, vm_id: str, silent=False, instance_type: str = 'vm'
+    ) -> Optional[List[Dict[str, Any]]]:
+        """Get VM by ID or Name.
+
+        This new implementation uses instance_type to distinct between
+        vm and template to either one instead of looking for both.
+
+        :param vm_id: VM ID (moref, uuid) or Name
+        :param silent: Silent mode
+        :param instance_type: Instance type
+        :return: Single object or None
+        """
         is_moref = validate_vm_moref('', '', vm_id)
         is_uuid = validate_uuid('', '', vm_id)
-        _LOGGING.debug(f'is_moref={is_moref}, is_uuid={is_uuid}')
+        _LOGGING.debug(f'{is_moref=} {is_uuid=} {vm_id=} {instance_type=}')
+        # instance_type lookup dictionary
+        lookup = {'vm': self.get_vms, 'template': self.get_templates}
+        try:
+            lookup_f = lookup[instance_type]
+        except KeyError:
+            raise click.BadArgumentUsage(
+                f'instance_type {instance_type} not supported'
+            )
+        # If it's a moref or uuid, then we can use the filter
         if is_moref or is_uuid:
             if is_moref:
                 attr = 'moref'
             else:
                 attr = 'uuid'
             filters = f'{attr},eq,{vm_id}'
-            v = self.get_vms(filter=filters)
+            v = lookup_f(filter=filters)
+            _LOGGING.debug(f'Using {lookup_f=} {attr=} {filters=}')
             if not v:
-                # try template
-                v = self.get_templates(filter=filters)
-                if not v:
-                    if silent:
-                        return None
-                    else:
-                        raise click.BadArgumentUsage(
-                            f'vm id {vm_id} could not be found'
-                        )
+                if silent:
+                    return None
+                else:
+                    raise click.BadArgumentUsage(
+                        f'{instance_type} id {vm_id} could not be found'
+                    )
             return v
         else:
             _LOGGING.debug(f'not a moref or uuid {vm_id}')
             # If it's a value error, then the string
             # is not a valid hex code for a UUID.
             # get vm by name
-            g_vms = self.get_vms(per_page=3000)
+            g_vms = lookup_f(per_page=3000)
             vm_id = vm_id.lower()
             v = list(filter(lambda i: vm_id in i['name'].lower(), g_vms))
             if not v:
-                # try templates:
-                g_tmpls = self.get_templates(per_page=2500)
-                v = list(filter(lambda i: vm_id in i['name'].lower(), g_tmpls))
                 if not v:
                     raise click.BadParameter(f'{vm_id} could not be found')
             v_count = len(v)
@@ -1751,7 +1766,7 @@ class Configuration(VssManager):
         """Ask assistant."""
         headers = {
             'Authorization': self.gpt_token,
-            'connection': 'keep-alive',
+            'Connection': 'keep-alive',
             'Content-Type': 'application/json',
         }
         retrieval_options = {
@@ -1763,6 +1778,7 @@ class Configuration(VssManager):
                 "document_set": None,
                 "time_cutoff": None,
                 "tags": [],
+                "user_file_ids": None,
             },
         }
         top_documents = []
@@ -1774,17 +1790,23 @@ class Configuration(VssManager):
         )
         # chat payload
         payload = {
+            "alternate_assistant_id": 2,
+            "chat_session_id": chat_id,
             "prompt_id": self._gpt_persona,
             "parent_message_id": None,
             "regenerate": False,
-            "alternate_assistant_id": 2,
             "full_doc": False,
+            "llm_override": {
+                "model_version": "gpt-4.1",
+                "model_provider": "Azure OpenAI - GPT 4.1",
+            },
             "search_doc_ids": [],
             "message": message,
             "file_descriptors": [],
-            "chat_session_id": chat_id,
             "retrieval_options": retrieval_options,
+            "use_agentic_search": False,
         }
+        _LOGGING.debug(f'User data payload {payload}')
         answer_text = ''
         with requests.post(
             f'{self.gpt_server}/api/chat/send-message',
@@ -1803,8 +1825,6 @@ class Configuration(VssManager):
                     if answer_piece is not None:
                         answer_text = answer_text + answer_piece
                         self.smooth_print(answer_piece)
-                    if answer_piece == "":
-                        break
         docs = []
         n = 1
         for doc in top_documents:
@@ -1817,7 +1837,6 @@ class Configuration(VssManager):
         answer_text = answer_text + '\n\n' + docs_text
         # clear console for formatting
         self.clear_console()
-
         from rich.console import Console
         from rich.markdown import Markdown